facade_pattern/requirements.txt
attrs==20.3.0
docopt==0.6.2
flake8==3.9.1
greenlet==1.0.0
mccabe==0.6.1
msgpack==1.0.2
neovim==0.3.1
neovim-remote==1.6.1
Pillow==8.2.0
psutil==5.8.0
pycodestyle==2.7.0
pyflakes==2.3.1
pynvim==0.4.3
python-xlib==0.29
ranger-fm==1.9.3
six==1.15.0
ueberzug==18.1.9
yapf==0.31.0
